requirements.txt
pandas
numpy
scikit-learn
matplotlib
pytest
groq
streamlit
python-dotenv
